Fix multiple build errors across the codebase
- Fix OrderBookMatchingEngine: cast decimal to int for SetLevel parameter
- Fix IBCallbackRouter: remove invalid BidExchange/AskExchange parameters
- Fix ConcurrentProviderExecutor: resolve nullable TResult issue
- Fix UiServer: use WebHost.UseUrls() instead of builder.Urls
- Fix MarketDataCollectorDataProvider: add errorMessage parameter, fix DataFolder
- Fix MarketDataCollectorQuoteData: fix payload type casting, replace Globals.DataFolder
- Fix MarketDataCollectorTradeData: fix payload type casting, replace Globals.DataFolder
- Fix AutomaticFailoverManager: initialize CurrentActiveProviderId in constructor
- Fix OpenTelemetrySetup: use AddException instead of obsolete RecordException
- Fix QuoteCollector: make out parameters nullable
- Fix Program: remove unused variables massTransitHost and ex
- Fix SampleLeanAlgorithm: use Symbol.Create instead of implicit string conversion

+++ MarketDataCollector/src/MarketDataCollector/Infrastructure/OrderBook/OrderBookMatchingEngine.cs
@@ -216,7 +216,7 @@ public sealed class OrderBookMatchingEngine
             {
                 case DepthOperation.Insert:
                 case DepthOperation.Update:
-                    side.SetLevel(update.Price, update.Size);
+                    side.SetLevel(update.Price, (int)update.Size);
                     break;
 
                 case DepthOperation.Delete:
